// apps/web/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ErrorDetail: {
            /** Field */
            field?: string | null;
            /** Location */
            location?: string | null;
            /** Message */
            message: string;
            /** Type */
            type?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}